chore: move import of docsearch style and change doc widget
