+++ src/me/components/DocSearchWidget.scss
@@ -198,6 +198,7 @@ $article-text: #d4d7dd;
 
   .DocSearch-Dropdown {
     padding-top: 16px !important;
+    max-height: 400px !important;
   }
 
   .DocSearch-Hits mark {
