+++ src/me/containers/MePage.scss
@@ -31,9 +31,9 @@
 
 .tutorials-list > li {
   font-size: 1.25em;
-  background-color: $g2-kevlar;
+  background-color: transparent;
   border-radius: $radius;
-  padding: $ix-marg-c;
+  padding: 8px;
   margin-bottom: $ix-marg-a;
 }
